viventium_v0_4/voice-gateway/requirements.txt
# === VIVENTIUM START ===
# Feature: LibreChat Voice Calls - Voice Gateway dependencies (pinned)
# Added: 2026-01-08
# === VIVENTIUM END ===

aiohttp==3.11.18
openai==2.5.0
livekit-agents==1.2.18
livekit-plugins-openai==1.2.18
livekit-plugins-silero==1.2.15; python_version < "3.13"
livekit-plugins-elevenlabs==1.2.18
## === VIVENTIUM START ===
## Feature: STT providers (AssemblyAI + local whisper.cpp)
## Added: 2026-01-11
## === VIVENTIUM END ===
livekit-plugins-assemblyai==1.2.18
numpy==2.1.3
pywhispercpp==1.4.0
